use long instead of int for request timeout type

+++ ethfirewall/src/main/java/tech/pegasys/ethfirewall/EthFirewallConfig.java
@@ -28,7 +28,7 @@ public interface EthFirewallConfig {
 
   Integer getDownstreamHttpPort();
 
-  Integer getDownstreamHttpRequestTimeout();
+  Long getDownstreamHttpRequestTimeout();
 
   String getHttpListenHost();
