+++ ethfirewall/src/main/java/tech/pegasys/ethfirewall/Runner.java
@@ -33,7 +33,7 @@ public class Runner {
   private TransactionSigner transactionSigner;
   private HttpClientOptions clientOptions;
   private HttpServerOptions serverOptions;
-  private Integer httpRequestTimeout;
+  private Long httpRequestTimeout;
   private Vertx vertx;
   private String deploymentId;
 
@@ -41,7 +41,7 @@ public class Runner {
       final TransactionSigner transactionSigner,
       final HttpClientOptions clientOptions,
       final HttpServerOptions serverOptions,
-      final Integer httpRequestTimeout) {
+      final Long httpRequestTimeout) {
     this.transactionSigner = transactionSigner;
     this.clientOptions = clientOptions;
     this.serverOptions = serverOptions;
