+++ ethfirewall/src/main/java/tech/pegasys/ethfirewall/jsonrpcproxy/JsonRpcHttpService.java
@@ -34,12 +34,12 @@ public class JsonRpcHttpService extends AbstractVerticle {
 
   private final RequestMapper requestHandlerMapper;
   private final HttpServerOptions serverOptions;
-  private final int httpRequestTimeout;
+  private final Long httpRequestTimeout;
   private HttpServer httpServer = null;
 
   public JsonRpcHttpService(
       final HttpServerOptions serverOptions,
-      final int httpRequestTimeout,
+      final Long httpRequestTimeout,
       final RequestMapper requestHandlerMapper) {
     this.serverOptions = serverOptions;
     this.httpRequestTimeout = httpRequestTimeout;
